+++ README.md
@@ -1,2 +1,2 @@
-# DatasetGenerationProcess
+# Dataset Generation Process 
 This repository stores the script that generates and exports the Measurements Dataset
